refactor: 도메인별 패키지 구조로 변경
- 클래스가 많아짐에 따라 도메인 기준으로 패키지 분리
- 공통으로 사용될 설정은 `global` 패키지로 관리
