+++ src/main/java/com/zeun/ddamap/station/service/StationApiService.java
@@ -1,6 +1,6 @@
-package com.zeun.ddamap.service;
+package com.zeun.ddamap.station.service;
 
-import com.zeun.ddamap.dto.StationApiResponse;
+import com.zeun.ddamap.station.dto.StationApiResponse;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
